ENH: Added some useful things to vim
We added the 80 char line at the the .vimrc so when we are programing we
know when to stop and make a new line. :)

Also implemented the last line when scrolling down to leave 2 lines at
the bottom it will make it easier to read the files.

We added the NERD Commenter to vim this will let you toggle the
commented and uncommented lines. with a simple <leader>c<space>

Miguel Kelly